Extract solc compile script
